+++ libs/components/config/projects/config/src/app-config-host.ts
@@ -15,6 +15,7 @@ const DEFAULTS: SkyuxConfigHost = {
 
 /**
  * Provides host configuration to components and applications.
+ * @deprecated Use `SkyAppConfig.skyux.host` instead.
  */
 @Injectable()
 export class SkyAppConfigHost {