+++ libs/components/config/projects/config/src/app-config-params.ts
@@ -1,15 +1,10 @@
 import {
-  Injectable,
-  Optional
+  Injectable
 } from '@angular/core';
 
 import {
   SkyuxConfigParams
 } from './config-params';
-
-import {
-  SkyAppParamsConfigArgs
-} from './params-config-args';
 
 const DEFAULTS = {
   params: {
@@ -26,25 +21,23 @@ const DEFAULTS = {
 };
 
 /**
- * @deprecated Use `SkyAppConfigParams` instead.
+ * @deprecated Use `SkyAppConfig.skyux.params` instead.
  */
-@Injectable({
-  providedIn: 'root'
-})
-export class SkyAppParamsConfig {
+@Injectable()
+export class SkyAppConfigParams {
 
   public get params(): SkyuxConfigParams {
-    return this._params || DEFAULTS.params;
+    return this._params;
   }
 
   private _params: SkyuxConfigParams;
 
-  constructor(
-    @Optional() args?: SkyAppParamsConfigArgs
-  ) {
+  public init(
+    params?: SkyuxConfigParams
+  ): void {
     this._params = {
       ...DEFAULTS.params,
-      ...args?.params || {}
+      ...params || {}
     };
   }