+++ libs/components/config/projects/config/src/app-config.module.spec.ts
@@ -14,10 +14,6 @@ import {
   SkyAppConfigParams
 } from './app-config-params';
 
-import {
-  SkyAppParamsConfig
-} from './params-config';
-
 describe('SkyAppConfigModule', () => {
 
   beforeEach(() => {
@@ -89,13 +85,9 @@ describe('SkyAppConfigModule.forRoot(config)', () => {
 
   it('should setup providers with config from forRoot args', () => {
     const hostConfig = TestBed.inject(SkyAppConfigHost);
-    const paramsConfig = TestBed.inject(SkyAppParamsConfig);
     const configParams = TestBed.inject(SkyAppConfigParams);
     expect(hostConfig.host.frameOptions).toEqual({
       blackbaud: true
-    });
-    expect(paramsConfig.params.foo).toEqual({
-      value: 'bar'
     });
     expect(configParams.params.foo).toEqual({
       value: 'bar'