+++ libs/components/config/projects/config/src/config.ts
@@ -76,6 +76,10 @@ export interface SkyuxConfigApp {
   title?: string;
 }
 
+export interface SkyuxConfigAuthSettings {
+  blackbaudEmployee?: boolean;
+}
+
 export interface SkyuxConfigHostBBCheckout {
   version: '2';
 }
@@ -103,12 +107,18 @@ export interface SkyuxConfigHost {
   url?: string;
 }
 
+export interface SkyuxConfigAnyAllSet {
+  any?: string[];
+  all?: string[];
+}
+
 export interface SkyuxConfig {
   $schema?: string;
   a11y?: SkyuxConfigA11y | boolean;
   app?: SkyuxConfigApp;
   appSettings?: any;
   auth?: boolean;
+  authSettings?: SkyuxConfigAuthSettings;
   codeCoverageThreshold?: 'none' | 'standard' | 'strict';
   command?: string;
   compileMode?: string;
@@ -126,10 +136,27 @@ export interface SkyuxConfig {
   params?: SkyuxConfigParams; // List of allowed params
   pipelineSettings?: any;
   plugins?: string[];
-  redirects?: any;
+  redirects?: {[from: string]: string};
   routes?: {
-    public?: any[],
-    referenced?: any[]
+    public?: {
+      title?: string
+      description?: string
+      global?: boolean,
+      route: string,
+      claims?: SkyuxConfigAnyAllSet,
+      entitlementType?: string,
+      entitlements?: SkyuxConfigAnyAllSet,
+      permissionId?: string,
+      permissions?: {
+        admin?: boolean,
+        legalEntityAdmin?: boolean,
+        permissionIds?: SkyuxConfigAnyAllSet
+      }
+    }[],
+    referenced?: {
+      app: string,
+      route: string
+    }[]
   };
   testSettings?: SkyuxConfigTestSettings;
   omnibar?: any;