+++ libs/components/config/projects/config/src/params-provider.spec.ts
@@ -3,30 +3,12 @@ import {
 } from './app-config-params';
 
 import {
-  SkyAppParamsConfig
-} from './params-config';
-
-import {
   SkyAppRuntimeConfigParamsProvider
 } from './params-provider';
 
 describe('SkyAppRuntimeConfigParamsProvider', () => {
 
   it('should return runtime params', () => {
-    const config = new SkyAppParamsConfig({
-      params: {
-        envid: {
-          value: 'foobar'
-        }
-      }
-    });
-
-    const provider = new SkyAppRuntimeConfigParamsProvider(config);
-
-    expect(provider.params.get('envid')).toEqual('foobar');
-  });
-
-  it('should use SkyAppConfigParams if defined', () => {
     const config = new SkyAppConfigParams();
     config.init({
       envid: {
@@ -34,7 +16,7 @@ describe('SkyAppRuntimeConfigParamsProvider', () => {
       }
     });
 
-    const provider = new SkyAppRuntimeConfigParamsProvider(undefined, config);
+    const provider = new SkyAppRuntimeConfigParamsProvider(config);
 
     expect(provider.params.get('envid')).toEqual('foobar');
   });