+++ libs/components/config/projects/config/src/params-provider.ts
@@ -8,15 +8,12 @@ import {
 } from './app-config-params';
 
 import {
-  SkyAppParamsConfig
-} from './params-config';
-
-import {
   SkyAppRuntimeConfigParams
 } from './params';
 
 /**
  * Provides methods to interact with runtime config query parameters.
+ * @deprecated Use `SkyAppConfig.runtime.params` instead.
  */
 @Injectable({
   providedIn: 'root'
@@ -30,12 +27,11 @@ export class SkyAppRuntimeConfigParamsProvider {
   private _params: SkyAppRuntimeConfigParams;
 
   constructor(
-    @Optional() paramsConfig?: SkyAppParamsConfig,
     @Optional() configParams?: SkyAppConfigParams
   ) {
     this._params = new SkyAppRuntimeConfigParams(
       window.location.href,
-      configParams?.params || paramsConfig?.params || {}
+      configParams?.params || {}
     );
   }