+++ libs/components/config/projects/config/src/public-api.ts
@@ -6,7 +6,5 @@ export * from './app-config.module';
 export * from './config-params';
 export * from './config';
 
-export * from './params-config-args';
-export * from './params-config';
 export * from './params-provider';
 export * from './params';